chore: rename trace-ai naming to bkn-trace
Follows the module rename in openbkn-ai/bkn-foundry#214. Renames the
internal src/trace-ai/ directory to src/bkn-trace/, moves
docs/product-specs/trace-ai.md to bkn-trace.md, and updates display
naming (Trace-AI / Trace AI / TraceAI -> BKN Trace) across docs, skill,
and command help text. CLI command surface (`openbkn trace ...`) and the
agent-observability API path are unchanged.
