+++ src/bkn-trace/schema-validate.ts
